feature/3: Renaming package for import path
