+++ graph/resultcell.go
@@ -1,4 +1,4 @@
-package rgraphquery
+package graph
 
 import (
 	"fmt"
